Moved Concat to matchers directory for API clarity
